Add I2CPacket class
This class encapsulates I2C packets, including all meta-data necessary for
bringing the package to the bus and handling the response.

The added sequence number is for house keeping in the requesting
application and allows de-coupling from packets to program flow. This is
especially helpful when packets are prioritized, i.e. the order is changed
by the sending entity.

This is a value class, which is immutable and thread-safe.