I2CPacket: introduce bus operations enum
Use an enum to specify the bus operation (based on SMBUS operations), add the
operation to each packet and allow to specify an i2c register. This allows
for flexible use of the I2C implementation outside the I3C context.

Add another constructor to create a "response packet", i.e. a packet with
matching meta-data but different data.